+++ packages/components/src/components.d.ts
@@ -9,7 +9,7 @@ import { BannerState, BreakpointCustomizable, ButtonType, ButtonVariant, FormSta
 import { ButtonGroupDirection } from "./components/layout/button-group/button-group-utils";
 import { FlexAlignContent, FlexAlignItems, FlexDirection, FlexInline, FlexJustifyContent, FlexWrap } from "./components/layout/flex/flex/flex-utils";
 import { FlexItemAlignSelf, FlexItemFlex, FlexItemGrow, FlexItemOffset, FlexItemShrink, FlexItemWidth } from "./components/layout/flex/flex-item/flex-item-utils";
-import { GridDirection } from "./components/layout/grid/grid/grid-utils";
+import { GridDirection, GridGutter, GridWrap } from "./components/layout/grid/grid/grid-utils";
 import { GridItemOffset, GridItemSize } from "./components/layout/grid/grid-item/grid-item-utils";
 import { HeadlineTag, HeadlineVariant } from "./components/basic/typography/headline/headline-utils";
 import { SpinnerSize } from "./components/feedback/spinner/spinner-utils";
@@ -192,7 +192,7 @@ export namespace Components {
          */
         "justifyContent"?: FlexJustifyContent;
         /**
-          * If set, overflowing elements will wrap to a new line.
+          * Handles wrapping behaviour of elements.
          */
         "wrap"?: FlexWrap;
     }
@@ -227,6 +227,14 @@ export namespace Components {
           * Defines the direction of the main and cross axis. The default "row" defines the main axis as horizontal left to right. Also defines the direction for specific breakpoints, like {base: "column", l: "row"}. You always need to provide a base value when doing this.
          */
         "direction"?: GridDirection;
+        /**
+          * Defines the gutter size for specific breakpoints. You always need to provide a base value when doing this.
+         */
+        "gutter"?: GridGutter;
+        /**
+          * Handles wrapping behaviour of elements.
+         */
+        "wrap"?: GridWrap;
     }
     interface PGridItem {
         /**
@@ -1122,7 +1130,7 @@ declare namespace LocalJSX {
          */
         "justifyContent"?: FlexJustifyContent;
         /**
-          * If set, overflowing elements will wrap to a new line.
+          * Handles wrapping behaviour of elements.
          */
         "wrap"?: FlexWrap;
     }
@@ -1157,6 +1165,14 @@ declare namespace LocalJSX {
           * Defines the direction of the main and cross axis. The default "row" defines the main axis as horizontal left to right. Also defines the direction for specific breakpoints, like {base: "column", l: "row"}. You always need to provide a base value when doing this.
          */
         "direction"?: GridDirection;
+        /**
+          * Defines the gutter size for specific breakpoints. You always need to provide a base value when doing this.
+         */
+        "gutter"?: GridGutter;
+        /**
+          * Handles wrapping behaviour of elements.
+         */
+        "wrap"?: GridWrap;
     }
     interface PGridItem {
         /**